Fix `deleteById(id)` and other test failures
use lodash to update the findBelongsTo which now returns an array of matches

ModelBaseClass: support promise-based observers

Allow the observer functions passed to `ModelBaseClass.observe`
to return a promise instead of calling the callback.

ModelBaseClass: promise mode for notifyObserversOf

Support both promise and callback styles in
ModelBaseClass.notifyObserversOf.

When there is no callback supplied, the method returns a promise that
is resolved (or rejected) with the result.

Add $now as shortcut default value for date property

Corrects improper stringification of nested values

Corrects property values which are arrays of objects (and possibly
nested), being updated in MySQL as [object Object], instead of the
actual stringified value