"Near" fix for Where clauses that contain and/or
The near filter on geo searches currently only searches a first level object: eg { "where": { "location": {"near": "30,-90"}}}.  For clauses that contain an "and" or an "or" clause, the geo filter does not pick up the operator, and you tend to get a SQL error back.  For instance, Postgres forms this abomination:  WHERE ("location"point($1,$2)) AND ("name" ~* $3).

This fix allows for queries with a format such as: "where":{"and":[{"location":{"near":"29,-90"}},{"name":{"regexp":"/sean/i"}}]}

It only handles one level of depth right now, as that's all we need, though I'm sure this could be made recursive to handle an infinite depth of and/or parameters.