Recurisve search for "near" operator in where clause
Implements recursive searching in "where" objects to find a "near"
operator at any depth.  It loops through every key in every object, and if
the value is an array, calls itself for every object in that array.  If
any of those objects contain an object with a "near" key, push that key
into the result set.  If more than one near object is returned, throw an
error, as multiple near objects are unsupported.