"Near" Fix For Where Clauses
The near filter on geo searches currently only searches a first level
object, eg { "where": { "location": {"near": "30,-90"}}}.  For
clauses that contain an "and" or an "or" clause, the geo filter
does not pick up the operator, and you tend to get a SQL error back.
For instance, Postgres forms this abomination:
WHERE ("location"point($1,$2)) AND ("name" ~* $3).

This fix allows for queries with a format such as:
"where":
{"and":[{"location":{"near":"29,-90"}},{"name":{"regexp":"/sean/i"}}]}

Additionally adds support for specifying multiple near filters. If a
point is within maxDistance of any of the near geopoints, it will be
included in the return result set.

"Near" Fix For Where Clauses

The near filter on geo searches currently only searches a first level
object, eg { "where": { "location": {"near": "30,-90"}}}.  For
clauses that contain an "and" or an "or" clause, the geo filter
does not pick up the operator, and you tend to get a SQL error back.
For instance, Postgres forms this abomination:
WHERE ("location"point($1,$2)) AND ("name" ~* $3).

This fix allows for queries with a format such as:
"where":
{"and":[{"location":{"near":"29,-90"}},{"name":{"regexp":"/sean/i"}}]}

Additionally adds support for specifying multiple near filters. If a
point is within maxDistance of any of the near geopoints, it will be
included in the return result set.

Fix syntax error

Fix merge conflicts