Correct syntax error in dao.js script for persist hook call.
They are a syntaxe error in updateAttributes and replaceById method for the Model.notifyObserversOf persist call.

The context past to the Model._forDB() method are not the context that are return by the persist hook.

There is a mixup between the variable 'ctx' returned by the persist hook and the variable 'context' that correspond to the above code block.